exports demo for begin project
