+++ src/main.js
@@ -2,6 +2,7 @@ import "./style.css";
 import javascriptLogo from "./javascript.svg";
 import viteLogo from "/vite.svg";
 import { setupCounter } from "./counter.js";
+import "./index.js";
 
 document.querySelector("#app").innerHTML = `
   <div>
